+++ tools/slang-test/options.cpp
@@ -126,13 +126,13 @@ static bool _isSubCommand(const char* arg)
             }
             optionsOut->binDir = *argCursor++;
         }
-        else if (strcmp(arg, "-use-test-proxy") == 0)
-        {
-            optionsOut->defaultSpawnType = SpawnType::UseProxy;
-        }
         else if (strcmp(arg, "-use-test-server") == 0)
         {
             optionsOut->defaultSpawnType = SpawnType::UseTestServer;
+        }
+        else if (strcmp(arg, "-use-fully-isolated-test-server") == 0)
+        {
+            optionsOut->defaultSpawnType = SpawnType::UseFullyIsolatedTestServer;
         }
         else if (strcmp(arg, "-v") == 0)
         {
